PLAT-1243 - Refactor share datalink logic
